tests: move tests out of line
