+++ src/math/Region.cpp
@@ -220,22 +220,3 @@ Vector2D Hyprutils::Math::CRegion::closestPoint(const Vector2D& vec) const {
 
     return leader;
 }
-
-#ifdef HU_UNIT_TESTS
-
-#include <gtest/gtest.h>
-
-TEST(Math, region) {
-    CRegion rg(CBox{{20, 20}, {40, 40}});
-
-    auto    extents = rg.getExtents();
-    EXPECT_EQ(extents.pos(), Vector2D(20, 20));
-    EXPECT_EQ(extents.size(), Vector2D(40, 40));
-
-    rg.scale(2);
-    extents = rg.getExtents();
-    EXPECT_EQ(extents.pos(), Vector2D(40, 40));
-    EXPECT_EQ(extents.size(), Vector2D(80, 80));
-}
-
-#endif
